Added features necessary to complete Assignment 5.

+++ src/main.c
@@ -1,12 +1,15 @@
 #include <stdio.h>
+#include <string.h>
 #include "../inc/customer.h"
 #include "../inc/employee.h"
 #include "../inc/ticket.h"
 #include "../inc/tools.h"
 
 int main() {
-	printf("Enter a number to select an option:\n1 - Enter a new customer\n2 - Enter a new employee\n3 - Enter a ticket\n4 -"
-			" View a ticket\n");
+	while(1){
+	printf("\nEnter a number to select an option:\n1 - Enter a new customer\n2 - Enter a new employee\n3 - Enter a ticket\n4 -"
+			" View a ticket\n5 - View all tickets\n6 - Delete a ticket\n7 - assign a ticket to a new employee\n"
+			"8 - unassign a ticket\n0 - exit the program\n");
 	int optionSelMain;
 	scanf("%d", &optionSelMain);
 	char catchBuffer[2];
@@ -17,10 +20,15 @@ int main() {
 		char ticketCreator[64];
 		printf("Enter customer name: ");
 		getline(ticketCreator, stdin);
-		struct customer* customer = createCustomer(ticketCreator);
-		writeCustomer(customer);
-		printf("New customer entered: %s\n", customer->name);
-		freeCustomer(customer);
+		printf("You are about to enter a new customer with the name: %s.\nDo you wish to continue? (y/n)\n", ticketCreator);
+		char confirm[2];
+		getline(confirm, stdin);
+		if (strcmp(confirm,"y")==0 || strcmp(confirm,"Y")==0){
+			struct customer* customer = createCustomer(ticketCreator);
+			writeCustomer(customer);
+			printf("New customer entered: %s\n", customer->name);
+			freeCustomer(customer);
+		}
 		break;}
 		
 		case 2:{
@@ -33,10 +41,18 @@ int main() {
 		printf("Enter the name of the employee's manager: ");
 		char empManager[64];
 		getline(empManager, stdin);
-		struct employee* employee = createEmployee(empName, empTitle, empManager);
-		writeEmployee(employee);
-		printf("New Employee Entered: %s\n", employee->name);
-		freeEmployee(employee);
+
+		printf("You are about to enter a new employee with the following attributes:\n"
+			"Name: %s\nTitle: %s\nManager: %s\n\nDo you wish to continue? (y/n)\n", empName, empTitle, empManager);
+
+		char confirm[2];
+		getline(confirm, stdin);
+		if (strcmp(confirm,"y")==0 || strcmp(confirm,"Y")==0){
+			struct employee* employee = createEmployee(empName, empTitle, empManager);
+			writeEmployee(employee);
+			printf("New Employee Entered: %s\n", employee->name);
+			freeEmployee(employee);
+		}
 		break;}
 		
 		case 3:{
@@ -61,11 +77,17 @@ int main() {
 		printf("Enter customer name: ");
 		char custName[64];
 		getline(custName, stdin);
-		struct customer* customer = loadCustomer(custName);
-		struct ticket* ticket = createTicket(subject, desc, prior, NULL, customer);
-		writeTicket(ticket);
-		printf("New Ticket Entered: %s\n", ticket->subjectLine);
-		freeTicket(ticket);
+		
+		printf("Please take a moment to review your entries. Do you want to enter the ticket? (y/n)\n");
+		char confirm[2];
+		getline(confirm, stdin);
+		if (strcmp(confirm, "y") == 0 || strcmp(confirm, "Y") == 0) {
+			struct customer* customer = loadCustomer(custName);
+			struct ticket* ticket = createTicket(subject, desc, prior, NULL, customer);
+			writeTicket(ticket);
+			printf("New Ticket Entered: %s\n", ticket->subjectLine);
+			freeTicket(ticket);
+		}
 		break;}
 
 		case 4: {
@@ -75,15 +97,56 @@ int main() {
 		struct ticket* ticket = loadTicket(sub);
 		char in[16];
 		strFromPriority(in, ticket->priority);
-		printf("Subject: %s Customer: %s %s\n%s\n", ticket->subjectLine, ticket->customer->name, in, ticket->ticketDesc);
+		printf("Subject: %s Customer: %s Priority: %s\n%s\n", ticket->subjectLine, ticket->customer->name, in, ticket->ticketDesc);
+		freeTicket(ticket);
+		break;}
+		
+		case 5: {
+		printAllTickets();
+		break;}
+
+		case 6: {
+		printf("Enter a ticket to delete: ");
+		char itemToDel[64];
+		getline(itemToDel, stdin);
+		printf("*******WARNING********\nDeleteing this ticket is permanent, do not continue unless you are sure you want to permanently delete this ticket! Are you sure you want to delete? (y/n)\n");
+		char confirm[2];
+		getline(confirm, stdin);
+		if (strcmp("y", confirm) == 0 || strcmp("Y", confirm) == 0) {	
+			deleteLine(itemToDel, "tickets");
 		}
 
-		case 5: {
-			printf("Enter a ticket to delete: ");
-			char itemToDel[64];
-			getline(itemToDel, stdin);
-			deleteLine(itemToDel, "tickets");
-			}
+		break;}
+
+		case 7: {
+		printf("Enter a ticket to reassign: ");
+		char subLine[64];
+		getline(subLine, stdin);
+		printf("Enter the new technician for this ticket: ");
+		char newEmp[64];
+		getline(newEmp, stdin);	
+		struct ticket* ticket = loadTicket(subLine);
+		struct employee* employee = loadEmployee(newEmp);
+
+		assignTicket(ticket, employee);
+		freeTicket(ticket);
+		freeEmployee(employee);
+		break;}
+
+		case 8: {
+		printf("Enter a ticket to unassign: ");
+		char subLine[64];
+		getline(subLine, stdin);
+		struct ticket* ticket = loadTicket(subLine);
+		assignTicket(ticket, NULL);
+		freeTicket(ticket);
+		break;}	
+
+		case 9: {
+		break;}
+
+		case 0: {
+		return 0;}
 	}
-	return 0;
+	}
 }
